requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dateparser==1.2.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
numpy==2.2.1
pandas==2.2.3
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pytz==2024.2
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
ta==0.11.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
